Extract shareable components from PredictionsStreamLive
